requirements.txt
streamlit==1.33.0
pandas==2.1.4
numpy==1.26.4
pyyaml==6.0.1
scikit-learn==1.4.2
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.14.1
reportlab==4.2.2
